% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ufmgthesis.R
\name{create_template}
\alias{create_template}
\title{Create a new ufmgthesis Quarto project from the package template}
\usage{
create_template(path = "ufmgthesis", open = TRUE, ...)
}
\arguments{
\item{path}{Path to create the project in (will be created if missing).}

\item{open}{If TRUE, opens the directory in RStudio (if available).}

\item{...}{further arguments passed to other methods.}
}
\description{
Create a new ufmgthesis Quarto project from the package template
}
